+++ frameworks/core/components_ng/pattern/rich_editor_drag/rich_editor_drag_pattern.cpp
@@ -86,7 +86,7 @@ RefPtr<FrameNode> RichEditorDragPattern::CreateDragNode(
     auto boxes = hostPattern->GetTextBoxes();
     for (const auto& child : imageChildren) {
         auto imageIndex = placeholderIndex[index];
-        if (imageIndex >= rectsForPlaceholders.size()) {
+        if (imageIndex >= static_cast<int32_t>(rectsForPlaceholders.size())) {
             break;
         }
         auto rect = rectsForPlaceholders.at(imageIndex);
